Functions moved from qd_database.py to qd_dissociate.py
