+++ src/qmflows/components/__init__.py
@@ -2,14 +2,14 @@ from .fde import (ADF3FDE_Result, Fragment, MFCC_Result, adf3fde, adf_fragmentsj
 from .operations import (find_first_job, select_max, select_min)
 
 from .reactivity import (Distance, Angle, Dihedral)
-from .qd_database import (read_database, compare_database, write_database, diss_list_to_pd)
+from .qd_database import (read_database, compare_database, write_database)
 from .qd_import_export import (read_mol, set_prop, create_dir, export_mol)
 from .qd_ams import (check_sys_var, qd_opt, ams_job_mopac_sp)
 from .qd_functions import (find_substructure, find_substructure_split, get_time,
                            rotate_ligand, merge_mol, qd_int, adf_connectivity, fix_h,
                            fix_carboxyl, update_coords)
 from .qd_ligand_opt import optimize_ligand
-from .qd_dissociate import get_topology_dict, dissociate_ligand
+from .qd_dissociate import (get_topology_dict, dissociate_ligand, diss_list_to_pd)
 
 __all__ = [
     'ADF3FDE_Result', 'Fragment', 'MFCC_Result', 'adf3fde', 'adf_fragmentsjob', 'mfcc',
@@ -17,7 +17,7 @@ __all__ = [
     'read_database', 'compare_database', 'write_database', 'diss_list_to_pd',
     'read_mol', 'set_prop', 'create_dir', 'export_mol',
     'check_sys_var', 'qd_opt', 'ams_job_mopac_sp',
-    'find_substructure', 'find_substructure_split', 'rotate_ligand',
+    'find_substructure', 'find_substructure_split', 'rotate_ligand', 'get_time',
     'merge_mol', 'qd_int', 'adf_connectivity', 'fix_h', 'fix_carboxyl', 'update_coords',
     'optimize_ligand',
-    'get_topology_dict', 'dissociate_ligand']
+    'get_topology_dict', 'dissociate_ligand', 'diss_list_to_pd']
