+++ src/qmflows/components/qd_database.py
@@ -117,46 +117,3 @@ def write_database(mol_list, database, path, mol_type='ligand'):
         path = os.path.join(path, name)
         database.to_excel(path + '.xlsx', sheet_name=sheet_name)
         database.to_json(path + '.json')
-
-
-def diss_list_to_pd(diss_list, residue_list, top_dict):
-    """
-    Converts specific entries from diss_list and residue_list to a pandas dataframe.
-    """
-    # Make all sublists of equal length
-    max_depth = len(residue_list[-1])
-    for res in residue_list:
-        len_res = len(res)
-        if len_res < max_depth:
-            res.extend([None for i in range(max_depth - len_res)])
-
-    gen1 = ((qd.properties.E, qd.properties.Eint, qd.properties.Estrain) for qd in diss_list)
-    dict1 = dict(zip(('E', 'Eint', 'Estrain'), zip(*gen1)))
-    gen2 = (itertools.chain(*(res, [top_dict[i] for i in res])) for res in residue_list)
-    keys = ('Residue numbers', 'Topology')
-    keys = [key + '_' + str(i+1) for key in keys for i in range(max_depth)]
-    dict1.update(dict(zip(keys, zip(*gen2))))
-    return pd.DataFrame(dict1)
-
-
-def average_energy(df):
-    """
-    Calculate the average of E, Eint & Estrain for the removal of each ligand with a given topology.
-    """
-    topology = [item for item in df.keys() if 'Topology' in item]
-    topology = np.array(df[topology])
-
-    vertice = np.array([(i, j, k) for i, j, k, top in
-                        zip(df['E'], df['Eint'], df['Estrain'], topology) if 'vertice' in top])
-    edge = np.array([(i, j, k) for i, j, k, top in
-                     zip(df['E'], df['Eint'], df['Estrain'], topology) if 'edge' in top])
-    face = np.array([(i, j, k) for i, j, k, top in
-                     zip(df['E'], df['Eint'], df['Estrain'], topology) if 'face' in top])
-
-    ret = pd.DataFrame({'vertice': np.mean(vertice, axis=0),
-                        'edge': np.mean(edge, axis=0),
-                        'face': np.mean(face, axis=0),
-                        'Energy': ('E', 'Eint', 'Estrain')})
-    ret.set_index('Energy')
-
-    return ret
